requirements.txt
playwright 
aiohttp 
beautifulsoup4 
pdfminer.six 
pymupdf 
pytesseract 
pillow 
sqlalchemy 
psycopg2-binary 
python-magic 
transformers 
sentence-transformers
faiss-cpu
tqdm
python-dotenv
playwright-asyncio
pdfplumber
lxml
webdriver-manager
selenium
requests
pandas
numpy
tabulate
retrying
openpyxl
pdf2image
pdfkit
reportlab
gunicorn
flask
flask-restful
flask-cors
flask-swagger-ui
marshmallow
marshmallow-sqlalchemy
flasgger
flask-jwt-extended
bcrypt
itsdangerous
celery
redis